Error: allEnemies not defined
Running the index file, resulted in a blank webpage.  CONSOLE in Chrome
Dev Tools revealed the error.

Created a temporary array declaration for allEnemies.  This allows the
IIFE Engine to run past this error.

Also corrected the spelling error for enemy in this file to meet style
convention and match engine.js name.