1st Enemy Sprite Starts Offscreen
Positioned 1 Enemy Sprite mostly offscreen on the left starting side.
Only a little bit of the front nose is showing.  Set the x value to -95
to do this.

This will give the appearance of the bug flowing onto the canvas as it
moves, instead of popping onto the first stone row tile.